Add PowerShell projection to SDK solution
