+++ src/api/Client/WinRT/powershell/Commands/Connections/CommandOpenMidiEndpointConnection.cs
