+++ src/api/Client/WinRT/powershell/Commands/Enumeration/CommandGetMidiEndpointGroups.cs
