+++ src/api/Client/WinRT/powershell/Types/MidiInfo.cs
