+++ src/api/Midi2-AppSDK.sln
@@ -149,6 +149,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Midi.Settings.Cor
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "midifixreg", "Client\WinRT\user-tools\midifixreg\midifixreg.vcxproj", "{3C2238C8-1087-40B5-9912-8BDBB6122183}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "PowerShell", "PowerShell", "{BDFAEF50-E40C-4FDC-8AB7-CFCD7DC7A0FE}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WindowsMidiServices", "Client\WinRT\powershell\WindowsMidiServices.csproj", "{68680138-9355-5A0D-4C91-1E60F8FF6AD4}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -999,6 +1003,30 @@ Global
 		{3C2238C8-1087-40B5-9912-8BDBB6122183}.Release|x64.Build.0 = Release|x64
 		{3C2238C8-1087-40B5-9912-8BDBB6122183}.Release|x86.ActiveCfg = Release|x64
 		{3C2238C8-1087-40B5-9912-8BDBB6122183}.Release|x86.Build.0 = Release|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Debug|Any CPU.Build.0 = Debug|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Debug|ARM64.ActiveCfg = Debug|ARM64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Debug|ARM64.Build.0 = Debug|ARM64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Debug|ARM64EC.ActiveCfg = Debug|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Debug|ARM64EC.Build.0 = Debug|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Debug|Win32.ActiveCfg = Debug|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Debug|Win32.Build.0 = Debug|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Debug|x64.ActiveCfg = Debug|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Debug|x64.Build.0 = Debug|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Debug|x86.ActiveCfg = Debug|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Debug|x86.Build.0 = Debug|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Release|Any CPU.ActiveCfg = Release|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Release|Any CPU.Build.0 = Release|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Release|ARM64.ActiveCfg = Release|ARM64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Release|ARM64.Build.0 = Release|ARM64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Release|ARM64EC.ActiveCfg = Release|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Release|ARM64EC.Build.0 = Release|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Release|Win32.ActiveCfg = Release|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Release|Win32.Build.0 = Release|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Release|x64.ActiveCfg = Release|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Release|x64.Build.0 = Release|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Release|x86.ActiveCfg = Release|x64
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4}.Release|x86.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1049,6 +1077,8 @@ Global
 		{C21D7C65-DAF5-8096-281F-5E76D3E53EF1} = {A6A9C5BD-1852-4464-BBCF-863A9D48A62A}
 		{234DEE5B-B930-38A3-62DB-11406CC1F4C8} = {A6A9C5BD-1852-4464-BBCF-863A9D48A62A}
 		{3C2238C8-1087-40B5-9912-8BDBB6122183} = {1C7F4144-9C66-4AE4-91AB-91DEF04407C2}
+		{BDFAEF50-E40C-4FDC-8AB7-CFCD7DC7A0FE} = {52E133B2-5F9B-4F79-9E62-070612BA9286}
+		{68680138-9355-5A0D-4C91-1E60F8FF6AD4} = {BDFAEF50-E40C-4FDC-8AB7-CFCD7DC7A0FE}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7F5A157F-9894-451D-B091-848921952166}
