--- src/app-sdk/projections/powershell/Commands/LoopbackEndpoints/CommandNewMidiLoopbackEndpointPair.cs
@@ -1,76 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License
-// ============================================================================
-// This is part of the Windows MIDI Services App API and should be used
-// in your Windows application via an official binary distribution.
-// Further information: https://aka.ms/midi
-// ============================================================================
-
-
-using System.Management.Automation;
-
-using Microsoft.Windows.Devices.Midi2.Reporting;
-using Microsoft.Windows.Devices.Midi2.Endpoints.Loopback;
-
-namespace WindowsMidiServices
-{
-
-    [Cmdlet(VerbsCommon.New, "MidiLoopbackEndpointPair")]
-    public class CommandNewMidiLoopbackEndpointPair : Cmdlet
-    {
-        [Parameter(Mandatory = true, Position = 0)]
-        public required string LoopbackBaseName
-        {
-            get; set;
-        }
-
-        [Parameter(Mandatory = true, Position = 1)]
-        public required string UniqueIdentifier
-        {
-            get; set;
-        }
-
-        [Parameter(Mandatory = false, Position = 2)]
-        public string? Description
-        {
-            get; set;
-        }
-
-        protected override void ProcessRecord()
-        {
-            try 
-            {
-                var sdkSessions = MidiReporting.GetActiveSessions();
-            }
-            catch
-            {
-                throw new Exception("No midi session found. Make sure to first run \"Start-MidiSession\"");
-            }
-
-            var loopbackA = new MidiLoopbackEndpointDefinition($"{LoopbackBaseName} (A)", UniqueIdentifier, Description);
-            var loopbackB = new MidiLoopbackEndpointDefinition($"{LoopbackBaseName} (B)", UniqueIdentifier, Description);
-
-            var guid = Guid.NewGuid();
-            var creationConfig = new MidiLoopbackEndpointCreationConfig(guid, loopbackA, loopbackB);
-            
-            try
-            {
-                var createdEndpoint = MidiLoopbackEndpointManager.CreateTransientLoopbackEndpoints(creationConfig);
-
-                WriteObject(createdEndpoint);
-            }
-            catch (Exception e)
-            {
-                ErrorRecord err = new ErrorRecord(
-                    e,
-                    "LoopbackCreationError",
-                    ErrorCategory.DeviceError,
-                    null);
-                WriteError(err);
-                return;
-            }
-        }
-    }
-
-
-}
