--- src/app-sdk/projections/powershell/WindowsMidiServices.psd1
@@ -1,134 +1,0 @@
-#
-# Module manifest for module 'WindowsMidiServices'
-#
-# Generated by: Pete Brown
-#
-# Generated on: 2/23/2025
-#
-
-@{
-
-# Script module or binary module file associated with this manifest.
-RootModule = 'WindowsMidiServices.dll'
-
-# Version number of this module.
-ModuleVersion = '0.0.1'
-
-# Supported PSEditions
-# CompatiblePSEditions = @()
-
-# ID used to uniquely identify this module
-GUID = 'd5a61bed-e196-4171-b6a4-6e6021be63db'
-
-# Author of this module
-Author = 'Pete Brown'
-
-# Company or vendor of this module
-CompanyName = 'Microsoft'
-
-# Copyright statement for this module
-Copyright = '(c) Microsoft. All rights reserved.'
-
-# Description of the functionality provided by this module
-Description = 'Integration with Windows MIDI Services'
-
-# Minimum version of the PowerShell engine required by this module
-PowerShellVersion = '7.4'
-
-# Name of the PowerShell host required by this module
-# PowerShellHostName = ''
-
-# Minimum version of the PowerShell host required by this module
-# PowerShellHostVersion = ''
-
-# Processor architecture (None, X86, Amd64) required by this module
-# ProcessorArchitecture = ''
-
-# Modules that must be imported into the global environment prior to importing this module
-# RequiredModules = @()
-
-# Assemblies that must be loaded prior to importing this module
-# RequiredAssemblies = @()
-
-# Script files (.ps1) that are run in the caller's environment prior to importing this module.
-# ScriptsToProcess = @()
-
-# Type files (.ps1xml) to be loaded when importing this module
-# TypesToProcess = @()
-
-# Format files (.ps1xml) to be loaded when importing this module
-# FormatsToProcess = @()
-
-# Modules to import as nested modules of the module specified in RootModule/ModuleToProcess
-# NestedModules = @()
-
-# Functions to export from this module, for best performance, do not use wildcards and do not delete the entry, use an empty array if there are no functions to export.
-FunctionsToExport = @()
-
-# Cmdlets to export from this module, for best performance, do not use wildcards and do not delete the entry, use an empty array if there are no cmdlets to export.
-CmdletsToExport = @(
-    "Start-Midi", "Stop-Midi", 
-    "Get-MidiSessionList", "Get-MidiEndpointDeviceInfo", "Get-MidiEndpointDeviceInfoList", "Get-MidiEndpointGroups", 
-    "Start-MidiSession", "Stop-MidiSession", 
-    "Open-MidiEndpointConnection", "Close-MidiEndpointConnection", 
-    "Send-MidiMessage", 
-    "Get-MidiMessageInfo",
-    "Start-MidiBeatClock", "Stop-MidiBeatClock",
-    "New-MidiLoopbackEndpointPair", "Remove-MidiLoopbackEndpointPair")
-
-# Variables to export from this module
-VariablesToExport = '*'
-
-# Aliases to export from this module, for best performance, do not use wildcards and do not delete the entry, use an empty array if there are no aliases to export.
-AliasesToExport = @()
-
-# DSC resources to export from this module
-# DscResourcesToExport = @()
-
-# List of all modules packaged with this module
-# ModuleList = @()
-
-# List of all files packaged with this module
-# FileList = @()
-
-# Private data to pass to the module specified in RootModule/ModuleToProcess. This may also contain a PSData hashtable with additional module metadata used by PowerShell.
-PrivateData = @{
-
-    PSData = @{
-
-        # Tags applied to this module. These help with module discovery in online galleries.
-        # Tags = @()
-
-        # A URL to the license for this module.
-        # LicenseUri = ''
-
-        # A URL to the main website for this project.
-        ProjectUri = 'https://aka.ms/midi'
-
-        # A URL to an icon representing this module.
-        # IconUri = ''
-
-        # ReleaseNotes of this module
-        # ReleaseNotes = ''
-
-        # Prerelease string of this module
-        # Prerelease = ''
-
-        # Flag to indicate whether the module requires explicit user acceptance for install/update/save
-        # RequireLicenseAcceptance = $false
-
-        # External dependent modules of this module
-        # ExternalModuleDependencies = @()
-
-    } # End of PSData hashtable
-
-} # End of PrivateData hashtable
-
-# HelpInfo URI of this module
-# HelpInfoURI = ''
-
-# Default prefix for commands exported from this module. Override the default prefix using Import-Module -Prefix.
-# DefaultCommandPrefix = ''
-
-}
-
