--- src/app-sdk/projections/powershell/midi-powershell.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.13.35723.152
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WindowsMidiServices", "WindowsMidiServices.csproj", "{037D9D12-DE3D-41D5-90AA-24A5931D2BE7}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|ARM64 = Debug|ARM64
-		Debug|x64 = Debug|x64
-		Release|Any CPU = Release|Any CPU
-		Release|ARM64 = Release|ARM64
-		Release|x64 = Release|x64
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{037D9D12-DE3D-41D5-90AA-24A5931D2BE7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{037D9D12-DE3D-41D5-90AA-24A5931D2BE7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{037D9D12-DE3D-41D5-90AA-24A5931D2BE7}.Debug|ARM64.ActiveCfg = Debug|ARM64
-		{037D9D12-DE3D-41D5-90AA-24A5931D2BE7}.Debug|ARM64.Build.0 = Debug|ARM64
-		{037D9D12-DE3D-41D5-90AA-24A5931D2BE7}.Debug|x64.ActiveCfg = Debug|x64
-		{037D9D12-DE3D-41D5-90AA-24A5931D2BE7}.Debug|x64.Build.0 = Debug|x64
-		{037D9D12-DE3D-41D5-90AA-24A5931D2BE7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{037D9D12-DE3D-41D5-90AA-24A5931D2BE7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{037D9D12-DE3D-41D5-90AA-24A5931D2BE7}.Release|ARM64.ActiveCfg = Release|ARM64
-		{037D9D12-DE3D-41D5-90AA-24A5931D2BE7}.Release|ARM64.Build.0 = Release|ARM64
-		{037D9D12-DE3D-41D5-90AA-24A5931D2BE7}.Release|x64.ActiveCfg = Release|x64
-		{037D9D12-DE3D-41D5-90AA-24A5931D2BE7}.Release|x64.Build.0 = Release|x64
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {423F67CD-6FC0-41F0-973C-BFCF69D89AA4}
-	EndGlobalSection
-EndGlobal
