+++ src/app-sdk/sdk-runtime-installer/main-bundle/Bundle.wxs
@@ -7,7 +7,7 @@
     <?define DependenciesSourceRootFolder=$(env.MIDI_REPO_ROOT)\build\dependencies ?>
   
     <?define StagingSourceRootFolder=$(env.MIDI_REPO_ROOT)\build\staging ?>
-    <?define StagingVersionIncludeFile=$(StagingSourceRootFolder)\version\BundleInfo.wxi ?>
+    <?define StagingVersionIncludeFile=$(StagingSourceRootFolder)\version\AppSdkVersion.wxi ?>
     <?include $(StagingVersionIncludeFile) ?>
 
     <!-- Set by build\build.ps1 -IncludePowerShellModule. The PowerShell projection has not been
